+++ common/src/main/java/com/aws/jverify/common/AnnotatedRange.java
@@ -1,0 +1,11 @@
+package com.aws.jverify.common;
+
+public class AnnotatedRange {
+    public final String annotation;
+    public final Range range;
+
+    public AnnotatedRange(String annotation, Range range) {
+        this.annotation = annotation;
+        this.range = range;
+    }
+}
