+++ common/src/main/java/com/aws/jverify/common/Range.java
@@ -1,0 +1,11 @@
+package com.aws.jverify.common;
+
+public class Range {
+    public final Position start;
+    public final Position end;
+
+    public Range(Position start, Position end) {
+        this.start = start;
+        this.end = end;
+    }
+}
